+++ src/core/colors/schemes/index.ts
@@ -1,3 +1,4 @@
 export * from './dark'
 export * from './soft-dark'
 export * from './light'
+export * from './aqua'